+++ alymova.ksenia/P4/reform.h
@@ -1,0 +1,8 @@
+#ifndef REFORM_H
+#define REFORM_H
+#include <cstddef>
+namespace alymova
+{
+  char* upper_string(const char* str, size_t end, char* str_res);
+}
+#endif
